chore: moved some loader files around
